VMAT DAO
Figured out problem with DAO: the leaf speed was defined as the
difference between positions, not absolute difference.  Then speed was
constrained to be between 0 and 6 cm/s, which forced unidirectional
motion, increasing the objective function.

Temporary fix: allow negative values.  Eventually, change definition of
function so that it considers the absolute difference.  Will also have
to change the Jacobian.