Bugfix in DijSampling
Old behaviour: if all bixels were in the core, the outer-core sampling
would still be attempted, resulting in an error message when attempting
to do the linspace (since the radDepthTail would be empty).

New behaviour: if all bixels are in the core, exit function with core
dose only.  Otherwise, proceed to rest of function.