Updates to where objectives and constraints are stored
Objectives and constraints are now stored as variables of the optimization problem. This takes away a lot of the redundant operations in the optimization function that loop over the cst otherwise.
Also changed the objective function to now calculate all objective function values in a separate function and then sum them up afterwards.
First changes for Pareto optimization were also added with the normalization options